+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/model/TrackTableModel.java
@@ -49,7 +49,7 @@ public final class TrackTableModel extends SwingEntityTableModel {
 	}
 
 	public void raisePriceOfSelected(BigDecimal increase) throws DatabaseException {
-		if (selectionModel().selectionNotEmpty().get()) {
+		if (selectionModel().selectionEmpty().not().get()) {
 			Collection<Long> trackIds = Entity.values(Track.ID, selectionModel().selectedItems().get());
 			Collection<Entity> result = connection()
 							.execute(Track.RAISE_PRICE, new RaisePriceParameters(trackIds, increase));