+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerTablePanel.java
@@ -55,7 +55,7 @@ public final class CustomerTablePanel extends EntityTablePanel {
 						.command(this::viewCustomerReport)
 						.name(BUNDLE.getString("customer_report"))
 						.smallIcon(FrameworkIcons.instance().print())
-						.enabled(tableModel().selectionModel().selectionNotEmpty())
+						.enabled(tableModel().selectionModel().selectionEmpty().not())
 						.build());
 	}