+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModel.java
@@ -443,11 +443,6 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		}
 
 		@Override
-		public StateObserver selectionNotEmpty() {
-			return selected.selectionEmpty.observer().not();
-		}
-
-		@Override
 		public Observer<?> selectionChanging() {
 			return selected.changing;
 		}